pyproject.toml
[project]
name = "gemini-code-fixer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-genai>=1.5.0",
]
